Rework the world.
Deleted all of the xxx-install.py apps and renamed them all to
xxx_install.py. They all now have a Class Installer with a install method.

Added test.sh which now tests that the install apps are at the expected
locations and that the versions are the expected values.

Changed gcc_install.py so all URL's are from gnu.org, the overseas
url's would sometime hangs.

Change utils.wget_extract to have a timeout to hopefully make it more
robust.