Add --crossDir option
Both binutils and gcc need to be placed into a directory separate from
'native' apps so with the prefix for ninja, meson is ~/opt it needs to
be ~/opt/xxx for binutils and gcc. I'm used the typical ~/opt/cross but
anthing else would do, it just can't be ~/opt other wise we get linker
errors.

The implementation of install.py assumes binutils and gcc are always cross
so I've renamed them in install.py and set DEFAULT_CROSS_DIR to "cross"
in binutils_install.py and gcc_install.py. This is good enough for now
but eventually we should be able to do cross or regular.