Simplify test.sh.
Pass app, the parameter for version and the code to get actual version
instead of being "tricky" in test_all.

Also added add_install_paths_to_org_path so that the array of
INSTALL_PATHS are added to the PATH with a ":" separator.

Finally used printGdbVer to also test we have the correct gdb.