Beginnings of creating an x86_64-pc-linux compiler.
In parseinstall call parse_known_args instead of parse_args so that
we can pass command line arguments to parseinstallargs.py that it
doesn't understand. It will return these unknown arguments in
self.unknownArgs.

In gcc_install.py we added support for the command line argument
"--extraGccConfigFlags". Multiple instances of --extraGccConfigFlags
maybe passed on the command line and all of them will be added to the
gcc ./configure command line.

Also, the gcc_install.__init__ constructor takes a new extraFlags
parameter which is an array of flags, you can see how it used in
install.py.

This is currently failing:

  $ ./install gcc-x86_64
  ...
  cc1: error: no include path in which to search for stdc-predef.h
  gcc_install.py: cwd=/home/wink/tmp/x86_64-pc-linux/gcc/build cmd= make install-gcc
  gcc_install.py: cwd=/home/wink/tmp/x86_64-pc-linux/gcc/build cmd= make all-target-libgcc -j 4
  configure: error: in `/home/wink/tmp/x86_64-pc-linux/gcc/build/x86_64-pc-linux/libgcc':
  configure: error: cannot compute suffix of object files: cannot compile
  See `config.log' for more details.
  make: *** [configure-target-libgcc] Error 1
  gcc_install.py: cwd=/home/wink/tmp/x86_64-pc-linux/gcc/build cmd= make install-target-libgcc
  make[1]: *** No rule to make target 'install'.  Stop.
  make: *** [install-target-libgcc] Error 2

So for now I'll leave it as is and come back to it in the future.