Use an explicit path to /usr/bin/python2 for bashPython2
On circleci the python3 environment can't create a python2 virtual env.
It complains it can't find a python2 and and that there are multiple
versions

 (venv-3.5.0) ubuntu@box31:~$ virtualenv --python=python2 venv2
 Running virtualenv with interpreter /home/ubuntu/.pyenv/shims/python2
 pyenv: python2: command not found

 The `python2' command exists in these Python versions:
  2.7.10
  2.7.3
  2.7.4
  2.7.5
  2.7.6
  2.7.7
  2.7.8
  2.7.9

So instead setting it to the preinstalled python2 with a specific
path seems to work when sshing to a circleci container,so we'll try that.