Get working on circleci.
Making the prerequisites isn't working so instead in circle.yml I apt-get
the the gmp, mpfr and mpc libraries.

I needed to reduce the cpu_count as the count is 32 on circleci and
running that many threads simultaneously caused the container to use to
> 4GB of memory and circleci aborted the build.

I turned on verbose output for `make all-gcc` as it didn't output anything
for > 10 minutes so circleci aborted the build. Originally I had to turn
off the verbosity (the default) because travis-ci aborted the build with
the log file was > 4M and was too big.

Finally, in utils.bashPython2 I had to use an explicit path to python2
because in circleci when the default python version is set to python3 it
uses virtualenv and pyenv. And when I tried to setup another virtualenv
inside the python3.5 virtualenv I got an error that python2 was ambiguous.
Using an explicit path resolves the problem although this will not work
if python2 is installed else where.